pyproject.toml
[project]
name = "dota-2-turbo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "requests>=2.32.4",
    "social-auth-app-django>=5.5.1",
]
